+++ pyalexaskill/AlexaBaseHandler.py
@@ -1,5 +1,6 @@
 import abc
 import logging
+import traceback
 
 
 class AlexaBaseHandler(object):
@@ -13,115 +14,10 @@ class AlexaBaseHandler(object):
 
     __metaclass__ = abc.ABCMeta
 
-    def __init__(self, app_id=None):
+    def __init__(self, app_id=None, log_level=logging.INFO):
         self.logger = logging.getLogger()
-        self.logger.setLevel(logging.INFO)
+        self.logger.setLevel(log_level)
         self.app_id = app_id
-
-    @abc.abstractmethod
-    def on_launch(self, launch_request, session):
-        """
-        Implement the LaunchRequest.  Called when the user issues a:
-        Alexa, open <invocation name>
-        :param launch_request:
-        :param session:
-        :return: the output of _build_response
-        """
-        pass
-
-    @abc.abstractmethod
-    def on_session_started(self, session_started_request, session):
-        pass
-
-    @abc.abstractmethod
-    def on_intent(self, intent_request, session):
-        """
-        Implement the IntentRequest
-        :param intent_request:
-        :param session:
-        :return: the output of _build_response
-        """
-        pass
-
-    @abc.abstractmethod
-    def on_help_intent(self, intent_request, session):
-        """
-        Implement the built in help intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Help Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_stop_intent(self, intent_request, session):
-        """
-        Implement the built in stop intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Stop Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_cancel_intent(self, intent_request, session):
-        """
-        Implement the built in cancel intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Cancel Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_no_intent(self, intent_request, session):
-        """
-        Implement the built in no (or negative) intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Answer No Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_yes_intent(self, intent_request, session):
-        """
-        Implement the built in yes (or positive) intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Answer Yes Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_repeat_intent(self, intent_request, session):
-        """
-        Implement the built in repeat intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Answer Repeat Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_start_over_intent(self, intent_request, session):
-        """
-        Implement the built in start over intent.
-        :param intent_request:
-        :param session:
-        :return:
-        """
-        raise ValueError("Start Over Intent was not implemented")
-
-    @abc.abstractmethod
-    def on_session_ended(self, session_end_request, session):
-        """
-        Implement the SessionEndRequest
-        :param session_end_request:
-        :param session:
-        :return: the output of _build_response
-        """
-        pass
 
     @abc.abstractmethod
     def on_processing_error(self, event, context, exc):
@@ -135,6 +31,165 @@ class AlexaBaseHandler(object):
         """
         pass
 
+    def on_intent(self, intent_request, session):
+        """
+        Implement the IntentRequest.  This is called if there is no specific
+        on_<intentname>_intent method.  This gives the implementer the choice
+        or writing specific on_intent methods, or calling a single method to
+        determine what needs to be called.
+        :param intent_request:
+        :param session:
+        :return: the output of _build_response
+        """
+        pass
+
+
+    def check_app_id(self, event):
+        """
+        Check the App id to make sure it is valid.
+        :param event:
+        :return: True - app id is valid, False - app id is invalid
+        """
+        valid_app_id = False
+        try:
+            if self.app_id and event['session']['application']['applicationId'] != self.app_id:
+                valid_app_id = False
+            else:
+                valid_app_id = True
+        except:
+            valid_app_id = False
+
+        return valid_app_id
+
+    def _handle_amazon_request(self, event, context):
+        """
+        Method dynamically calls AMAZON built in requests.  For example, with the
+        new Audio Playback, there call requests of the form:
+        AudioPlayer.PlaybackStarted
+        AudioPlayer.PlaybackFinished
+        To keep this generic, this method will dynamically call method of the form:
+
+        A.B
+        on_a_b(event, context)
+        e.g:
+        AudioPlayer.PlaybackStarted will be a method
+        on_audioplayer_playbackstarted(..)
+
+        :param event:
+        :param context:
+        :return:speechlet_response, directive, None
+        """
+        response = None
+        self.logger.debug("_handle_amazon_request: event: {0}".format(event))
+        request_type = event['request']['type']
+        self.logger.debug("_handle_amazon_request: {0}".format(request_type))
+        if request_type:
+            parts = request_type.split(".")
+            if len(parts) == 1:
+                request_type_method_name = "on_{0}".format(parts[0].lower())
+            elif len(parts) == 2:
+                request_type_method_name = "on_{0}_{1}".format(parts[0].lower(), parts[1].lower())
+            else:
+                raise NotImplementedError("Unexpected request type: {0}".format(request_type))
+
+            self.logger.debug("_handle_amazon_request: {0}".format(request_type_method_name))
+            if hasattr(self, request_type_method_name):
+                try:
+                    response = getattr(self, request_type_method_name)(event, context)
+                except:
+                    self.logger.error("Traceback Exception {0}".format(traceback.format_exc()))
+                    self.logger.error("ERROR: _handle_amazon_request: {0}".format(request_type_method_name))
+                    raise
+            else:
+                # not every request is required to be implemented - particularly for the
+                # playbackcontroller requests
+                self.logger.warn("_handle_amazon_request: {0} method not found".format(request_type_method_name))
+                #raise NotImplementedError("No method with name: {0} exists in class".format(request_type_method_name))
+
+        return response
+
+    def _handle_amazon_intent(self, event, context):
+        """
+        Method dynamically calls AMAZON built in intents.
+        For example, for the AMAZON.YesIntent, this method will dynamically call
+        a method of the form:
+        on_yes_intent(intent_request, session)
+
+        This allows for this base class to be extensible to handle the new
+        Amazon Alexa Streaming Playback support without having to specially
+        look for the streaming intent names.
+        :param event:
+        :param context:
+        :return: speechlet_response
+        """
+        response = None
+
+        intent_name = self._get_intent_name(event['request'])
+        if intent_name is not None and intent_name.startswith("AMAZON."):
+            intent_method_name = "on_{0}_intent".format(intent_name.split(".")[1].replace("Intent","").lower())
+            self.logger.debug("_handle_amazon_intent: {0}".format(intent_method_name))
+
+            if hasattr(self, intent_method_name):
+                try:
+                    response = getattr(self, intent_method_name)(event['request'], event['session'])
+                except:
+                    self.logger.error("Traceback Exception {0}".format(traceback.format_exc()))
+                    self.logger.error("ERROR: _handle_amazon_intent: {0}".format(intent_method_name))
+                    raise
+
+            else:
+                raise NotImplementedError("No method with name: {0} exists in class".format(intent_method_name))
+
+        return response
+
+    def _handle_custom_intent(self, event, context):
+        """
+        Method dynamically calls customeintents by taking the intent name,
+        lower casing it, and prepending the name with on_
+
+        For example, for the MyCustomIntent, this method will dynamically call
+        a method of the form:
+        on_mycustomintent(intent_request, session)
+
+        If the class has not method with that name, it will look for a
+        method with the name, 'on_intent'.
+
+        If no method with that name exists then an error will be thrown.
+
+        :param event:
+        :param context:
+        :return: speechlet_response
+        """
+        response = None
+
+        intent_name = self._get_intent_name(event['request'])
+        if intent_name is not None:
+            intent_method_name = "on_{0}_intent".format(intent_name.lower())
+            self.logger.debug("_handle_custom_intent: {0}".format(intent_method_name))
+
+            if hasattr(self, intent_method_name):
+                try:
+                    response = getattr(self, intent_method_name)(event['request'], event['session'])
+                except:
+                    self.logger.error("Traceback Exception {0}".format(traceback.format_exc()))
+                    self.logger.error("ERROR: _handle_amazon_intent: {0}".format(intent_method_name))
+                    raise
+
+            elif hasattr(self, 'on_intent'):
+                try:
+                    response = getattr(self, 'on_intent')(event['request'], event['session'])
+                except:
+                    self.logger.error("Traceback Exception {0}".format(traceback.format_exc()))
+                    self.logger.error("ERROR: _handle_custom_intent: {0}".format(intent_method_name))
+                    raise
+            else:
+                raise NotImplementedError("No method with name: {0} exists in class".format(intent_method_name))
+
+        return response
+
+    #--------------------------------------------------------
+    #--------------- Main Processing Entry Point  -----------
+    #--------------------------------------------------------
     def process_request(self, event, context):
         """
         Helper method to process the input Alexa request and
@@ -143,40 +198,48 @@ class AlexaBaseHandler(object):
         :param context:
         :return: response from the on_ handler
         """
+        self.logger.debug("process_request: event: {0}".format(event))
+
+        try:
+            request_type = event['request']['type']
+            self.logger.debug("event[request][type]: {0}".format(request_type))
+        except:
+            request_type = None
+
+        try:
+            new_session = event['session']['new']
+        except:
+            new_session = False
+
         # if its a new session, run the new session code
         try:
-            if(self.app_id and event['session']['application']['applicationId'] != self.app_id):
-                raise ValueError("Invalid Application ID")
+            if not self.check_app_id(event):
+                raise NotImplementedError("Invalid Application ID")
 
             response = None
-            if event['session']['new']:
-                self.on_session_started({'requestId': event['request']['requestId']}, event['session'])
+            if new_session is not False and hasattr(self, 'on_session_started'):
+                # then it is a new session, and the concrete class has an on_session_started
+                # callback
+                getattr(self, 'on_session_started')(event['request'], event['session'])
 
-                # regardless of whether its new, handle the request type
-            if event['request']['type'] == "LaunchRequest":
-                response = self.on_launch(event['request'], event['session'])
-            elif event['request']['type'] == "IntentRequest":
+            # regardless of whether its new, handle the request type
+            if request_type == "IntentRequest":
+                # Only handle IntentRequest here... all others in the else block
                 intent_name = self._get_intent_name(event['request'])
-                if intent_name == "AMAZON.HelpIntent":
-                    response = self.on_help_intent(event['request'],event['session'])
-                elif intent_name == "AMAZON.StopIntent":
-                    response = self.on_stop_intent(event['request'], event['session'])
-                elif intent_name == "AMAZON.CancelIntent":
-                    response = self.on_cancel_intent(event['request'], event['session'])
-                elif intent_name == "AMAZON.NoIntent":
-                    response = self.on_no_intent(event['request'], event['session'])
-                elif intent_name == "AMAZON.RepeatIntent":
-                    response = self.on_repeat_intent(event['request'], event['session'])
-                elif intent_name == "AMAZON.StartOverIntent":
-                    response = self.on_start_over_intent(event['request'], event['session'])
-                elif intent_name == "AMAZON.YesIntent":
-                    response = self.on_yes_intent(event['request'], event['session'])
+                if intent_name is not None and intent_name.startswith("AMAZON."):
+                    response = self._handle_amazon_intent(event, context)
                 else:
-                    response = self.on_intent(event['request'], event['session'])
-            elif event['request']['type'] == "SessionEndedRequest":
-                response = self.on_session_ended(event['request'], event['session'])
+                    # this is a user specific intent, so let the users concrete
+                    # implementation handle it.
+                    response = self._handle_custom_intent(event, context)
+            else:
+                # LaunchRequest, SessionEndedRequest, AudioPlayer requests, etc
+                # are handled here.
+                self.logger.debug("Calling _handle_amazon_request")
+                response = self._handle_amazon_request(event, context)
 
         except Exception as exc:
+            self.logger.error("Error in process_request: {0}".format(traceback.format_exc()))
             self.logger.error(exc.message)
             response = self.on_processing_error(event, context, exc)