+++ CHANGELOG.rst
@@ -119,3 +119,22 @@ Changes:
 * update create_aws_lambda.py to allow for files in directories below the root directory
 and they will be copied into a corresponding deployment directory.
 
+0.2.0 (2016-10-10)
+------------------
+
+Changes:
+~~~~~~~~
+* many of the abstract methods were removed, because with the AudioPlayer capability, also
+  came additional intents and request types.  Instead of trying to keep up with these, I have
+  opted for dynamically calling methods based on a convention for the intent and request
+  names.
+* added audio intent handling and generically call amazon intents, custom intents and audio requests
+* BREAKING CHANGE: on_start_over_intent needs to be renamed to on_startover_intent
+* BREAKING CHANGE: on_launch needs to be renamed to on_launchrequest and are passed event, context
+* BREAKING CHANGE: on_session_ended needs to be renamed to on_sessionendedrequest and are passed event, context
+* added ability to include requirements-test.txt which create_aws_lambda.py will install as:
+    pip install -i https://testpypi.python.org/pypi <requirements line> -t <deployment_dir>
+   to allow for test packages to be added to a Lambda function zip file.
+* added log level specification to ctor
+* added on_invalid_response_request abstract method
+* added create_alexa_audio_handler.py to create a starter template for audio applications